Correct printing of queries without names
It would incorrectly print unnamed queries in shorthand format by just
testing the name for being empty. This is not entirely correct as also
variable definitions and directives need to be empty.

Made it more robust by having a member on OperationDefinition that can
tell whether a query can be represented in shorthand format.

Also add a couple of test cases related to this.